src/test/javascript/spec/app/entities/work-effort-inventory-produced/work-effort-inventory-produced.service.spec.ts
import { TestBed, getTestBed } from '@angular/core/testing';
import { HttpClientTestingModule, HttpTestingController } from '@angular/common/http/testing';
import { WorkEffortInventoryProducedService } from 'app/entities/work-effort-inventory-produced/work-effort-inventory-produced.service';
import { IWorkEffortInventoryProduced, WorkEffortInventoryProduced } from 'app/shared/model/work-effort-inventory-produced.model';

describe('Service Tests', () => {
  describe('WorkEffortInventoryProduced Service', () => {
    let injector: TestBed;
    let service: WorkEffortInventoryProducedService;
    let httpMock: HttpTestingController;
    let elemDefault: IWorkEffortInventoryProduced;
    let expectedResult: IWorkEffortInventoryProduced | IWorkEffortInventoryProduced[] | boolean | null;

    beforeEach(() => {
      TestBed.configureTestingModule({
        imports: [HttpClientTestingModule],
      });
      expectedResult = null;
      injector = getTestBed();
      service = injector.get(WorkEffortInventoryProducedService);
      httpMock = injector.get(HttpTestingController);

      elemDefault = new WorkEffortInventoryProduced(0, 0);
    });

    describe('Service methods', () => {
      it('should find an element', () => {
        const returnedFromService = Object.assign({}, elemDefault);

        service.find(123).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'GET' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(elemDefault);
      });

      it('should create a WorkEffortInventoryProduced', () => {
        const returnedFromService = Object.assign(
          {
            id: 0,
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.create(new WorkEffortInventoryProduced()).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'POST' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(expected);
      });

      it('should update a WorkEffortInventoryProduced', () => {
        const returnedFromService = Object.assign(
          {
            quantity: 1,
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.update(expected).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'PUT' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(expected);
      });

      it('should return a list of WorkEffortInventoryProduced', () => {
        const returnedFromService = Object.assign(
          {
            quantity: 1,
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.query().subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'GET' });
        req.flush([returnedFromService]);
        httpMock.verify();
        expect(expectedResult).toContainEqual(expected);
      });

      it('should delete a WorkEffortInventoryProduced', () => {
        service.delete(123).subscribe(resp => (expectedResult = resp.ok));

        const req = httpMock.expectOne({ method: 'DELETE' });
        req.flush({ status: 200 });
        expect(expectedResult);
      });
    });

    afterEach(() => {
      httpMock.verify();
    });
  });
});
